requirements.txt
django-filter==22.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
requests==2.26.0
PyJWT==2.4.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
asgiref==3.3.2
Django==3.2
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1